Install SaltStack dependencies using pip
SaltStack has stopped packaging some dependencies for 32 bits [0]. This
means we cannot use the bootstrap script to install the required
dependencies while keeping 32 bit compatibility.

To solve this, the required Python packages are first installed using
pip, then the bootstrap script is called with the -b option (assume
dependencies are already installed).

The pip dependencies are frozen to a specific version to avoid future
changes as much as possible.